tests/wpt/tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-case2.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<meta http-equiv="Content-Security-Policy" content="trusted-types 'nONe'">
<body>
<script>
  // CSP3 is defined in terms of RFC 5234 grammars, which states that terminal
  // symbols are case-insensitive. So 'none' should be parsed the
  // same as 'NONE'.
  //
  // Ref:
  // - https://www.w3.org/TR/CSP3/#framework-infrastructure
  // - https://www.rfc-editor.org/rfc/rfc5234#section-2.3
  // - See also: https://github.com/w3c/webappsec-csp/issues/236
  test(t => {
    assert_throws_js(TypeError, _ => {
      window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    });
  }, "Cannot create a policy when CSP lists 'none'");
</script>
